requirements.txt
google-generativeai
google-cloud-logging
pandas
pydantic>=2.0.0
fastapi
uvicorn
python-dotenv
python-multipart